test: reorganize tests to mirror conversion package
