+++ .github/workflows/benchmark.yml
@@ -124,12 +124,12 @@ jobs:
         timeout-minutes: 20
         run: |
           start_time=$(date +%s.%N)
-          poetry run pytest tests/test_analysis/test_geojson.py::test_csv_to_geojson_big_file -v -s --durations=0 --input_file="$TEST_GEOCSV_FILENAME"
+          poetry run pytest tests/test_analysis/test_analysis_csv.py::test_analyse_csv_big_file -v -s --durations=0
           end_time=$(date +%s.%N)
           execution_time=$(echo "scale=2; ($end_time - $start_time) / 1" | bc -l)
           COMMIT_SHA="${{ github.sha }}"
           SHORT_SHA="${COMMIT_SHA:0:7}"
-          echo "$(date -u +"%Y-%m-%dT%H:%M:%SZ"),test_csv_to_geojson_big_file,${{ env.TEST_GEOCSV_FILE }},github,$execution_time,${{ github.actor }},$SHORT_SHA,${{ matrix.resource-class }},$(nproc),$(free -m | awk 'NR==2{printf "%.0f", $2}'),${{ env.PYTHON_VERSION }}" >> .benchmarks/benchmarks.csv
+          echo "$(date -u +"%Y-%m-%dT%H:%M:%SZ"),test_analyse_csv_big_file,${{ env.TEST_GEOCSV_FILE }},github,$execution_time,${{ github.actor }},$SHORT_SHA,${{ matrix.resource-class }},$(nproc),$(free -m | awk 'NR==2{printf "%.0f", $2}'),${{ env.PYTHON_VERSION }}" >> .benchmarks/benchmarks.csv
 
       - name: Commit CSV to GeoJSON benchmark results
         run: |
@@ -160,7 +160,7 @@ jobs:
         timeout-minutes: 20
         run: |
           start_time=$(date +%s.%N)
-          poetry run pytest tests/test_analysis/test_geojson.py::test_geojson_to_pmtiles_big_file -v -s --durations=0 --input_file="$TEST_GEOJSON_FILENAME" -k "test_geojson_to_pmtiles_big_file"
+          poetry run pytest tests/test_conversion/test_geojson_to_pmtiles.py::test_geojson_to_pmtiles_big_file -v -s --durations=0 --input_file="$TEST_GEOJSON_FILENAME" -k "test_geojson_to_pmtiles_big_file"
           end_time=$(date +%s.%N)
           execution_time=$(echo "scale=2; ($end_time - $start_time) / 1" | bc -l)
           COMMIT_SHA="${{ github.sha }}"
